requirements.txt
streamlit
snowflake-connector-python
pandas
PyPDF2
scikit-learn
python-dotenv
nltk
numpy

google-generativeai
watchdog
openpyxl
